% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ExampleDD.R
\docType{data}
\name{ExampleDD}
\alias{ExampleDD}
\title{Objeto de clase DD.}
\format{Objeto de clase \code{DD}. Su slot \code{Data} tiene 52 filas y 6
columnas.}
\usage{
data(ExampleDD)
}
\description{
\code{ExampleDD} es un objeto de clase \linkS4class{DD}.
}
\details{
Las columnas del slot Data del objeto son las siguientes:
\itemize{
  \item Variable. Identificador de variable estadística.
  \item Sort. Rol de la variable (IDQual, NonIDQual, IDDD).
  \item Class. Clase en el lenguaje R de la variable.
  \item Qual1. Primer calificador de la variable.
  \item Qual2. Segundo calificador de la variable.
  \item Qual3. Tercer calificador de la variable.
}
}

