% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleDM.R
\docType{data}
\name{ExampleDM}
\alias{ExampleDM}
\title{Conjunto de datos con estructura matricial.}
\format{\code{data.table} con 29499 filas y 59 columnas.}
\usage{
data(ExampleDM)
}
\description{
\code{ExampleDM} es una \link{data.table} con estructura de datos matricial, 
con los datos correspondientes a un periodo no especificado de la encuesta de
 IASS.
}
\details{
Los nombres de las columnas son aquellos de las variables en el repositorio, 
anexando los calificadores con un guión bajo cuando sea necesario (ver 
\code{DD} para más información acerca de sus significados).
}

