% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleQ.R
\docType{data}
\name{ExampleQ}
\alias{ExampleQ}
\title{Objeto de clase StQ.}
\format{Objeto de clase \code{StQ}. Su slot \code{Data} tiene 1710942 filas
 y 6 columnas.}
\usage{
data(ExampleQ)
}
\description{
\code{ExampleQ} es un objeto de clase \linkS4class{StQ} cuyo slot
\code{Data} es una \code{\link{data.table}} con estructura par clave-valor
con los datos de la encuesta de IASS correspondientes a un periodo no 
especificado (el mismo periodo utilizado para el objeto \code{ExampleDM}).
}
\details{
Las columnas del slot Data del objeto son las siguientes:
\itemize{
  \item NOrden. Primer calificador (identificador de unidad).
  \item CCAA. Segundo calificador (de variable).
  \item EsRemuner. Tercer calificador (de variable).
  \item TipoRem. Cuarto calificador (de variable).
  \item IDDD. Nombres de las variable.
  \item Valor. Valores de las variable (por defecto, de clase \code{character}).
}
}

