% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleStQ.R
\docType{data}
\name{ExampleStQ}
\alias{ExampleStQ}
\title{Object of class \code{\link[StQ]{StQ}}.}
\format{Object of class \code{\link[StQ]{StQ}}. Its slot \code{Data} has 354458 rows and 4 
columns.}
\usage{
data(ExampleStQ)
}
\description{
\code{ExampleStQ} is an object of class \code{\link[StQ]{StQ}} containing severely
perturbed anonymised data from a real official statistical operation.
}
