% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FirstLine.R
\docType{methods}
\name{FirstLine}
\alias{FirstLine}
\alias{FirstLine,StQ-method}
\alias{FirstLine,data.table-method}
\title{Generate the first line of a file}
\usage{
FirstLine(object)

\S4method{FirstLine}{data.table}(object)

\S4method{FirstLine}{StQ}(object)
}
\arguments{
\item{object}{Objeto de clase \code{\linkS4class{StQ}}.}
}
\value{
\code{Character} \code{vector} of length 1 with the variables in the specified \code{StQ}
object and theirs corresponding formats, that is, with the register desing of the file. Besides,
it includes \code{NIV} as first variable, with value specified as input param, and \code{M} as
final variable corresponding to the maximum register length in the file.
}
\description{
\code{FirstLine} returns a character \code{vector} of length 1 with the first line
to appear in the file for the input object.

This function takes an object of class \linkS4class{StQ} as an input parameter and returns the
first line to appear in the corresponding file when the input object is to be written (see
function \link{WriteRepoFile}). The information contained in this first line is essentially the
schema of the file.
}
\examples{
data(ExampleStQ)
FirstLine(ExampleStQ)
}
\seealso{
\link{ReadRepoFile}, \link{WriteRepoFile}
}

