% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateWeights.R
\name{GenerateWeights}
\alias{GenerateWeights}
\title{Generates StQ object of aggregate/index weights}
\usage{
GenerateWeights(SASFilesNames, DD, SurveyName)
}
\arguments{
\item{SASFilesNames}{A named character vector with the corresponding SAS files of weightings. The 
name of each component must be the name of the corresponding Excel sheet of weightings.}

\item{DD}{Object of class \linkS4class{DD} with the content of the file \code{DD} of definitions 
and properties of every variable.}

\item{SurveyName}{Character vector of length 1 with the name of the survey.}
}
\value{
\linkS4class{StQ} object with the content of the SAS files of weights corresponding to 
the SheetNamesPond.
}
\description{
\code{GenerateWeights} returns a \linkS4class{StQ} with data from the SAS files of 
weighting corresponding to the specified SAS files.

This function reads the SAS files of weightings corresponding to the Excel sheets specified as 
input parameter \code{SheetNamesPond} and returns its content in a \linkS4class{StQ} object. 

This correspondence is specified through the respective component from \code{VarNameCorresp} 
slot of the \code{DD} input object, which is a \link{data.table} with the content of a sheet of 
the Excel file, and \code{DD}, which is an object of class \linkS4class{DD} containing the DD  
file with the definition and properties of every variable. 

The object \code{DD} is naturally obtained from the original \code{DD} file as output of 
functions \code{\link{RepoDDToDD}} or \code{\link{xmlToDD}}.
}
\examples{
\dontrun{
# Param SASWeighFiles must be as following:
SASWeightFiles <- c('C:/Users/Administrador/Desktop/pondcarama14.sas7bdat')
names(SASWeightFiles) <- 'PondCARama'
SASWeightFiles

# We assume data created previosly:
Weightss <- GenerateWeights(SASFilesNames, DD, SurveyName = 'IASS')
}

}

