% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NextRepoFile.R
\docType{methods}
\name{NextRepoFile}
\alias{NextRepoFile}
\alias{NextRepoFile,FileName-method}
\title{Create the name of the next version of a file}
\usage{
NextRepoFile(Periods, Path, FileType)

\S4method{NextRepoFile}{FileName}(Periods, Path, FileType)
}
\arguments{
\item{Periods}{Character vector with time periods included in the file names.}

\item{Path}{Character vector with the path of the search directory.}

\item{FileType}{Character vector with the sort (FF, FD, FG, DD, ...) of the files.}
}
\value{
It returns a character vector of length equal to the length of \code{Periods} with the
full names of the next version of each input file name.
}
\description{
\code{RepoNextFileVerNumber} returns the name of the next version of a file in a
directory.

This function takes every file corresponding to the input file sort \code{FileType} for the input
periods \code{Periods} in the input directory \code{Path} and generates for each of them the name
of their next version.
}
\examples{
\dontrun{
RepoTopn(Path, Type)
}
}

