% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseFileName.R
\name{ParseFileName}
\alias{ParseFileName}
\title{Produce a vector of file names parsed out of a pattern and a sequence of time period names}
\usage{
ParseFileName(FileNames, TimePeriods)
}
\arguments{
\item{FileNames}{Character vector with the names of the files to parse.}

\item{TimePeriods}{Character vector with the time periods.}
}
\value{
Return a vector similar to the input vector \code{FileNames} with the period codes parsed
 according to the input \code{TimePeriods}.
}
\description{
This function parses each component of a vector of file names with a pattern filling
 out the corresponding codes with the input time periods.
}
\examples{
\dontrun{
ParseFileName(c('N:/UDMTD/UDTMDCOM/DepSel.Repositorio/E30183/DataSAS/Ruta_dat/dat[mm][aa].sas7bdat'), c('MM122015', 'MM012016', 'MM022016'))
}

}

