% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParseUnitName.R
\name{ParseUnitName}
\alias{ParseUnitName}
\title{Produce a vector of file names parsed out of a pattern and a sequence of time period names}
\usage{
ParseUnitName(UnitNames, TimePeriods)
}
\arguments{
\item{UnitNames}{Character vector with the names of the files to parse.}

\item{TimePeriods}{Character vector with the time periods.}
}
\value{
Return a vector with each component of the input vector \code{UnitNames} parsed according
 to the input parameter \code{TimePeriods}.
}
\description{
This function parses each component of a vector of file names with a pattern filling
 out the corresponding codes with the input time periods.
}
\examples{
ParseUnitName(c('dat[mm][aa].sas7bdat'), c('MM122015', 'MM012016', 'MM022016'))
ParseUnitName(c('dat[mm][aa].sas7bdat', 'in[aaaa][mm].txt'), c('MM122015', 'MM012016', 'MM022016'))

}

