% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadOldRepoFile.R
\name{ReadOldRepoFile}
\alias{ReadOldRepoFile}
\title{Write a file with a key-value}
\usage{
ReadOldRepoFile(FileName, language = "SP", perl = FALSE)
}
\arguments{
\item{FileName}{Character vector of length 1 with the name of the file to read. The file will be 
read from the working directory (see \link[base]{getwd}) unless the full path is specified.}

\item{language}{Character vector of length 1 with the language of the file to read. The values 
allowed are: 'SP' (Spanish) and 'EN' (English), being the default value 'SP'.}

\item{perl}{Logical vector of length 1 indicating whether Perl is installed in the system or not.}
}
\value{
\linkS4class{data.table} with all data from the input file.
}
\description{
\code{ReadOldRepoFile} returns a \linkS4class{data.table} with the content of the 
file corresponding to the input name.
}
\examples{
\dontrun{
#We assume that the key-value ASCII file \\code{E30183.FF_V1.MM032014.D_1} is in the administrator 
desktop (change accordingly otherwise): 
RepoName <- 'C:/Users/Administrador/Desktop/E30183.FF_V1.MM032014.D_1'
Example.kv <- ReadOldRepoFile(RepoName)
str(Example.kv)
}

}
\seealso{
\code{\link{WriteRepoFile}}
}
