% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSASFile.R
\name{ReadSASFile}
\alias{ReadSASFile}
\title{Read a SAS file renaming variables as specified in input parameters}
\usage{
ReadSASFile(SASFileName, Exceldf, DD, DDslot = "MicroData")
}
\arguments{
\item{SASFileName}{Character vector of length 1 with name of the file. The 
file will be read from the working directory (see \link[base]{getwd}) unless 
the full path is specified.}

\item{Exceldf}{\code{\link{data.frame}} con el contenido del fichero Excel 
de equivalencias entre los nombres de las variables de la encuesta asignados 
por las distintas unidades del INE implicadas.}

\item{DD}{Object of class \linkS4class{DD} with the content of the file 
\code{DD} of definitions and properties of every variable.}
}
\value{
\linkS4class{data.table} with the contents of the \code{DD} file, 
with statistical units in rows and variables as columns.
}
\description{
\code{ReadSASFile} returns a \linkS4class{data.table} with all 
data from the input SAS file.

This function reads the SAS file specified as input parameter 
\code{SASFileName} and returns its content in a \linkS4class{data.table} with
statistical units in rows and variables in columns. The names of the 
variables are assigned according to the Excel file with variable names 
correspondence for each statistical operation.

This correspondence is specified with the input parameters \code{Exceldf}, 
which is a \link{data.frame} with the content of a sheet of the Excel file, 
and \code{DD}, which is a \link{data.frame} with the content of the DD file 
containing the definition and properties of every variable. 

The \link{data.frame} \code{DD} is naturally obtained from the original 
\code{DD} file as output of functions \code{\link{RepoDDToDD}} or
\code{\link{xmlToDD}}.
}
\examples{
# We assume that the SAS file \\code{MM032014.sas7bdat} is in the 
administrator desktop:
SASName <- 'C:/Users/Administrador/Desktop/MM032014.sas7bdat'
# We assume data created previosly:
data(XLS)
data(RepoDD)
data(VNC)
DD <- RepoDDToDD(RepoDD, VNC)
Example.DM <- ReadSASFile(SASName, XLS, DD)

}
\seealso{
\link{RepoDDToDD}, \link{ReadRepoFile}, \link{WriteRepoFile}
}

