% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSASWeights.R
\name{ReadSASWeights}
\alias{ReadSASWeights}
\title{Read a SAS file of aggregate/index weights renaming variables according to the DD input object}
\usage{
ReadSASWeights(SASFileName, DD, DDslot = "MicroData", VNCName = DDslot)
}
\arguments{
\item{SASFileName}{Character vector of length 1 with the name of the file. The file will be read 
from the working directory (see \link[base]{getwd}) unless the full path is specified.}

\item{DD}{Object of class \linkS4class{DD} with the definitions and properties of every variable.}

\item{DDslot}{Character vector of length 1 with the name of the DD slot in which data to be read 
are defined. Its default value is \code{MicroData}.}

\item{VNCName}{Character vector of length 1 with the name of the component of the list of slot 
\linkS4class{VarNameCorresp} in which data to be read are defined. Its default value is the value
 of the \code{DDslot} input parameter.}
}
\value{
\linkS4class{data.table} with the contents of the \code{DD} file, with statistical units 
in rows and variables as columns.
}
\description{
\code{ReadSASWeights} returns a \linkS4class{data.table} with all data from the 
input SAS file of aggregate/index weights.

This function reads the SAS file of weights specified as input parameter, \code{SASFileName}, 
and returns its content in a \linkS4class{data.table} with statistical units in rows and variables 
in columns. The names of the variables are assigned according to the Excel file with variable 
names correspondence for each statistical operation.

This correspondence is stablished with the component \code{VNCName} specified as input param, of 
the \linkS4class{VarNameCorresp} slot of the \linkS4class{DD} input object. That component is a 
\linkS4class{data.table} with the content of a sheet of the Excel file

The object \code{DD} is naturally obtained from the original \code{DD} file as output of 
functions \code{\link{RepoDDToDD}} or \code{\link{xmlToDD}}.
}
\examples{
\dontrun{
# We assume data created previosly in the SAS file SASName:
Weights <- ReadSASWeights(SASName, DD, VNCName = 'PondCARama')
}

}
\seealso{
\link{RepoDDToDD}, \link{ReadRepoFile}, \link{WriteRepoFile}
}

