% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadXLSFile.R
\name{ReadXLSFile}
\alias{ReadXLSFile}
\title{Read a XLS file renaming variables as specified in input parameters}
\usage{
ReadXLSFile(XLSFileName, DD)
}
\arguments{
\item{XLSFileName}{Character vector of length 1 with name of the file. The file will be read from 
the working directory (see \link[base]{getwd}) unless the full path is specified.}

\item{DD}{Object of class \linkS4class{DD} with the content of the file \code{DD} of definitions 
and properties of every variable.}
}
\value{
\linkS4class{data.table} with the contents of the \code{DD} file, with statistical units 
in rows and variables as columns.
}
\examples{
\dontrun{
# We assume that the SAS file \\code{mm201604.xlsx} is in the 
#administrator desktop:
XLSFileName <- 'C:/Users/Administrador/Desktop/mm201604.xlsx'
# We assume data created previosly:
data(ExampleDD)
Example.DM <- ReadXLSFile(XLSFileName, ExampleDD)
}

}
\seealso{
\link{RepoDDToDD}, \link{ReadRepoFile}, \link{WriteRepoFile}
}

