% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoDDToDD.R
\name{RepoDDToDD}
\alias{RepoDDToDD}
\title{Produce an object of class \linkS4class{DD} from a DD file}
\usage{
RepoDDToDD(RepoDD, VNC)
}
\arguments{
\item{RepoDD}{\link{data.frame} with the content of the file \code{DD}.}

\item{VNC}{Object of class \linkS4class{VarNameCorresp}}
}
\value{
Object of class \linkS4class{DD}.
}
\description{
This function is a constructor for the class \linkS4class{DD} 
using the contents of a DD file in the original ASCII format.

\code{RepoDDToDD} transforms a \link{data.frame} with the content of a DD 
file with the original ASCII fixed-width columnwise format into an object of 
class \linkS4class{DD}. 

This function internally builds a \linkS4class{data.table} with columns
 \code{Variable}, \code{Sort}, \code{Class} and \code{Qual1} to 
 \code{Qual\emph{q}}. The column \code{Variable} contains the names of all 
 variables, both questionnaire variables and metadata. This internal 
 \linkS4class{data.table} is then used to initialize a \linkS4class{DD} 
 object.

The column \code{Sort} takes values \code{'IDQual'}, \code{'NonIDQual'} or 
\code{'IDDD'}, for statistical unit qualifiers, variable name qualifiers and
variable names, respectively.

The column \code{Class} specifies the class of the variable and takes values
\code{numeric} or \code{character}. 

The columns \code{Qual1} to \code{Qual\emph{q}} contain the names of the 
qualifiers of every variable name (row).
}
\examples{
# We assume that the xlsx file \\code{NombresVariables.E30053.xlsx} whit the
# appropriate structure
library(data.table)   
ExcelName <- 'R:/MacrosSAS.Doc/NombresVariables.E30053.xlsx'
SheetNames <- c('Datos', paste0(c('poicn0.', 'poicn1.', 'poicn2m.','ponca.',
                 'pond_divi.', 'ponderas.'),substring(MesInicial,5,8)))
VNC <- RepoXLSToVNC(ExcelName, SheetNames)
RepoDDToDD(RepoDD, VNC)

}

