% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoFileToTXT.R
\name{RepoFileToTXT}
\alias{RepoFileToTXT}
\title{@title Write a key-value pair txt file.}
\usage{
RepoFileToTXT(RepoPath, StQList, FileTyPe, RutaEscritura)
}
\description{
@description \code{RepoFileToTXT} writes the key-value objects found in the input parameter 
 StQList in fixed-width column txt files.
}
\details{
This method reads the files found with the input parameter StQList and writes their content as a
 fixed-width column txt file with key-value pair structure with the same name as files read, but
 extension .txt.

@param RepoPath Character vector of length 1 with the path of the repository from which files are
 to be read.
 
 @param StQList Object of class \linkS4class{StQList}.
 
 @param FileType Character vector of length 1 with the type of the file to be written (FI, FF, FG 
 or FD).

@param RutaEscritura Character vector of length 1 with the path where files will be written. It
 must not be another txt file of the same periods which are being considered.
}
\examples{
\dontrun{
 StQList <- RepoFileToStQList('Z:/', 'XXXXXX', 'C:/', 'E30183', 'MM112014', 'MM122014', 'FF')
 RepoFileToTXT('Z:/', StQList, 'FF', 'C:/')
}
      
}

