% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoNextFileVerNumber.R
\name{RepoNextFileVerNumber}
\alias{RepoNextFileVerNumber}
\title{Create the name of the next version of a file}
\usage{
RepoNextFileVerNumber(Periods, Path, FileType, Base, DorP = "D")
}
\arguments{
\item{Periods}{Character vector with time periods included in the file names.}

\item{Path}{Character vector with the path of the search directory.}

\item{FileType}{Character vector with the sort (FF, FD, FG, DD, ...) of the files.}

\item{Base}{Character vector of length 1 with the year ('aaaa') of the base to which data are
referred. If it has no sense, "Base" is a empty character vector.}

\item{DorP}{Character vector with the type of version (D or P) of the files.}
}
\value{
It returns a character vector of length equal to the length of \code{Periods} with the 
next version of each input file name.
}
\description{
\code{RepoNextFileVerNumber} returns the name of the next version of a file in a 
directory.

This function takes every file corresponding to the input file sort \code{FileType} for the input 
periods \code{Periods} in the input directory \code{Path} and generates for each of them the name 
of their next version.
}
\examples{
\dontrun{
RepoNextFileVerNumber(c('MM012016'), 'C:/Repo/E30183', 'FF', '2015')
}

}
