% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoPeriodRange.R
\name{RepoPeriodRange}
\alias{RepoPeriodRange}
\title{Obtain the time periods of a given input file type in a given directory}
\usage{
RepoPeriodRange(Path, Type, Last = out[length(out)])
}
\arguments{
\item{Path}{Character vector of length 1 specifying the path of the searching directory.}

\item{Type}{Character vector of length 1 specifying the type of file whose time periods are 
queried. It can take values \code{FF}, \code{FG}, \code{FD}, \code{FL} or \code{FT}.}

\item{Last}{Character vector of length 1 specifying the last time period. By default it is the 
last of all of them.}
}
\value{
Character vector with as many components as time periods present in the input directory 
up to the last one specified (included).
}
\description{
\code{RepoPeriodRange} returns the time periods of files of a given input file type 
\code{Type} (FF, FD, FG, FL, FT, FI, FP, FA) in an input directory \code{Path} up to a last given
 input time period \code{Last} (included). By default, it returns all time periods in the queried
  directory unless the input parameter \code{Last} is specified.
}
\examples{
\dontrun{
RepoPeriodRange(Path = 'T:/E30183', Type = 'FF', Last = 'MM102014')
}

}

