% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RepoTopn.R
\name{RepoTopn}
\alias{RepoTopn}
\title{Obtain index number of the last version of a file in a directory}
\usage{
RepoTopn(Path, FileNameString)
}
\arguments{
\item{Path}{Character vector of length 1 specifying the path of the searching directory.}

\item{FileNameString}{Character vector of length 1 specifying the string of the file name whose
last version is to be searched.}
}
\value{
Integer vector of length 1 with the index number of the last version of the file.
}
\description{
\code{RepoTopn} returns the index number of the last version of a file identified
with a file name in a given directory.

This function searches in the input directory \code{Path} for the last version of a file with
file name containing the string \code{FileNameString} and returns the corresponding index number.
}
\examples{
\dontrun{
RepoTopn('R:/E30183', 'FF_V1.MM122014')
}
}

