% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoXLSToDD.R
\name{RepoXLSToDD}
\alias{RepoXLSToDD}
\title{Produce an object \linkS4class{DD} from an Excel file}
\usage{
RepoXLSToDD(ExcelName)
}
\arguments{
\item{ExcelName}{Character vector of length 1 with the name of the file to read. The file will be 
read from the working directory (see \link[base]{getwd}) unless the full path is specified.}
}
\value{
This function returns an object of class \linkS4class{DD}.
}
\description{
This function builds an object \linkS4class{DD} using the contents of an xlsx file.
\code{RepoXLSToDD} transforms the content of an xlsx file into a DD file.
}
\examples{
# We assume that the xlsx file ExampleXLS.NombresVariables.xlsx with the appropriate structure is
# in the working directory (change accordingly otherwise):

\dontrun{
ExcelName <- 'T:/E30163/E30163.NombresVariables_V1.xlsx'
DD <- RepoXLSToDD(ExcelName)
show(DD)
}

}

