% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateXLS.R
\name{ValidateXLS}
\alias{ValidateXLS}
\title{Validate xls file to create DD}
\usage{
ValidateXLS(ExcelName)
}
\arguments{
\item{ExcelName}{Character vector of length 1 with the name of the file to read.The file will be 
read from the working directory (see \link[base]{getwd}) unless the full path is specified.}
}
\description{
This function validate the xls file it will be used to create the XLM file DD.

\code{ValidateXLS} checks the xls file to be consistent to create the XML file DD.
}
\examples{
\dontrun{
ExcelName <- 'N:/UDMTD/UDTMDCOM/DepSel.Repositorio/E30163/E30163.NombresVariables.xlsx'
Validate(ExcelName)
}

}

