% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FirstLine.R
\docType{methods}
\name{FirstLine}
\alias{FirstLine}
\alias{FirstLine,StQ-method}
\alias{FirstLine,data.table-method}
\title{Generate the first line of a file}
\usage{
FirstLine(object)

\S4method{FirstLine}{data.table}(object)

\S4method{FirstLine}{StQ}(object)
}
\arguments{
\item{object}{Objeto de clase \code{\linkS4class{StQ}}.}

\item{NIV}{\code{\link{vector}} de longitud 1 con el número de niveles en el
fichero. 
Corresponde al número total de calificadores que aparecen en el fichero.Por 
defecto toma el valor \code{NIV} = \code{1}.}
}
\value{
\code{Vector} de tipo \code{character} de longitud 1 con las
variables que aparecen en el objeto \code{StQ} especificado y sus formatos
correspondientes, es decir, con el diseño de registro del fichero. Incluye
además como primera variable NIV, con el valor especificado en el input, y
una última, \code{M}, que corresponde a la máxima longitud de registro en el
fichero.
}
\description{
\code{FirstLine} returns a character \code{vector} of length 1
with the first line to appear in the file for the input object.

This function takes an object of class \linkS4class{StQ} as an input 
parameter and returns the first line to appear in the corresponding file when
the input object is to be written (see function \link{WriteRepoFile}). The 
information contained in this first line is essentially the schema of the 
file.
}
\examples{
data(ExampleQ)
FirstLine(ExampleQ)

}
\seealso{
\link{ReadRepoFile}, \link{WriteRepoFile}
}

