% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.validXLSX.R
\name{is.validXLSX}
\alias{is.validXLSX}
\title{Check validity of an xlsx file to create an DD (Data Dictionary).}
\usage{
is.validXLSX(ExcelName, verbose = FALSE)
}
\arguments{
\item{ExcelName}{Character vector of length 1 with the name of the file to read.}

\item{verbose}{\code{TRUE} or \code{FALSE} (default) to request verbose mode.}
}
\value{
\code{TRUE} or \code{FALSE}.
}
\description{
This function checks the validity of the xlsx file used as a GUI to create the data 
dictionary with XML format.
}
\details{
The input xlsx file is a naive user interface to generate a data dictionary in XML 
format with the specification of each single key-value pair in the dictionary. The function 
performs a sequence of checks over each sheet of the input xlsx file. These checks are necessary 
(though not sufficient) for the construction of the XML file with the function 
\code{\link{RepoXLSToDD}}.
}
\examples{
path <- system.file('extdata', package = 'RepoReadWrite')
ExcelName <- 'E30103.NombresVariables_V1.xlsx'
is.validXLSX(file.path(path, ExcelName), verbose = TRUE)

}
