% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadRepoFile.R
\name{ReadRepoFile}
\alias{ReadRepoFile}
\title{Write a file with a key-value}
\usage{
ReadRepoFile(FileName, DD)
}
\arguments{
\item{FileName}{Character vector of length 1 with the name of the file to read. The file will be 
read from the working directory (see \link[base]{getwd}) unless the full path is specified.}

\item{DD}{Object of class \linkS4class{DD} with the definition and characteristics of the data 
contained in the file to read.}
}
\value{
\linkS4class{data.table} with all data from the read file.
}
\description{
\code{ReadRepoFile} returns a \linkS4class{data.table} with the content of the file 
corresponding to the input name.
}
\examples{
\dontrun{
#We assume that the key-value ASCII file \\code{E30183.FF_V1.MM032014.D_1} is in the administrator 
desktop (change accordingly otherwise): 
RepoName <- 'C:/Users/Administrador/Desktop/E30183.FF_V1.MM032014.D_1'
Example.kv <- ReadRepoFile(RepoName)
str(Example.kv)
}

}
\seealso{
\code{\link{ReadSASFile}}, \code{\link{ReadXLSFile}}, \code{\link{WriteRepoFile}}
}

