<<<<<<< HEAD
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoDDToDD.R
\name{RepoDDToDD}
\alias{RepoDDToDD}
\title{Produce an object of class \linkS4class{DD} from a DD file}
\usage{
RepoDDToDD(RepoDD, VNC, DDslot = "MicroData")
}
\arguments{
\item{RepoDD}{\link{data.frame} with the content of the file \code{DD}.}

\item{VNC}{Object of class \linkS4class{VarNameCorresp}.}

\item{DDslot}{Character vector of length 1 with the name of DD slot in which
transformation will be made. Its default value is \code{MicroData}.}
}
\value{
Object of class \linkS4class{DD}.
}
\description{
This function is a constructor for the class \linkS4class{DD} 
using the contents of a DD file in the original ASCII format.

\code{RepoDDToDD} transforms a \link{data.frame} with the content of a DD 
file with the original ASCII fixed-width columnwise format into an object of 
class \linkS4class{DD}. 

This function internally builds a \linkS4class{data.table} with columns
 \code{Variable}, \code{Sort}, \code{Class} and \code{Qual1} to 
 \code{Qual\emph{q}}. The column \code{Variable} contains the names of all 
 variables, both questionnaire variables and metadata. This internal 
 \linkS4class{data.table} is then used to initialize a \linkS4class{DD} 
 object.

The column \code{Sort} takes values \code{'IDQual'}, \code{'NonIDQual'} or 
\code{'IDDD'}, for statistical unit qualifiers, variable name qualifiers and
variable names, respectively.

The column \code{Class} specifies the class of the variable and takes values
\code{numeric} or \code{character}. 

The columns \code{Qual1} to \code{Qual\emph{q}} contain the names of the 
qualifiers of every variable name (row).
}
\examples{
# An example with data created previosly:
library(data.table)
data(RepoDD)
data(VNC)
RepoDDToDD(RepoDD, VNC)

}

||||||| merged common ancestors
=======
% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RepoDDToDD.R
\name{RepoDDToDD}
\alias{RepoDDToDD}
\title{Transforma el fichero DD del repositorio a su formato para R}
\usage{
RepoDDToDD(RepoDD)
}
\arguments{
\item{RepoDD}{\code{data.frame} con el contenido del fichero
\code{DD} de definición de datos ubicado en el repositorio de microdatos de
la encuesta en su formato para R.}
}
\value{
\code{data.table} con los datos del fichero \code{DD}
en el formato adecuado para su inclusión como slot \code{DD} de un objeto de
clase \code{StQ}.
}
\description{
Esta función es un constructor para la clase \linkS4class{DD} a partir del
contenido del fichero DD correspondiente en el repositorio.
}
\details{
\code{RepoDDToDD} transforma un \code{\linkS4class{data.frame}} con el contenido
del fichero DD de definición de datos ubicado en el repositorio de microdatos
de la encuesta, en un \code{\linkS4class{data.table}} con el formato requerido para
su inclusión como slot \code{DD} de un objeto de clase \code{\linkS4class{StQ}}.

Esta función toma el contenido del \code{data.frame} de entrada y
construye un \code{data.table} con las columnas \code{Variable},
\code{Sort}, \code{Class} and \code{Qual\code{1}} to \code{Qual\code{q}}.

La columna \code{Variable} contiene los nombres de todas las variables, tanto
de las del cuestionario como de las creadas durante el proceso.

La columna \code{Sort} toma los valores \code{'IDQual'}, \code{'NonIDQual'},
e \code{'IDDD'}, para calificadores de identificación de las unidades
estadísticas, para calificadores de nombres de variables y para nombres de
variables otras variables en el cuestionario.

La columna \code{Class} especifica la clase de la variable, y toma los valores
\code{numeric} o \code{character}.

Las columnas \code{Qual\code{1}} a \code{Qual\code{q}} contienen los nombres de los
calificadores de cada variable (fila).

Se transforma el fichero de definición de datos para que su tratamiento desde
el punto de vista de la programación sea más eficiente.
}
\examples{
data(RepoDD)
RepoDDToDD(RepoDD)
}

>>>>>>> 8f2c18abd559e81f9d5a50ddf707317c405ee90f
