% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoDDToDD.R
\name{RepoDDToDD}
\alias{RepoDDToDD}
\title{Produce an object of class \linkS4class{DD} from a xml file}
\usage{
RepoDDToDD(FileName, VNC, DDslot = NULL)
}
\arguments{
\item{FileName}{Character vector of length 1 with the name of the file to read. The file will be 
read from the working directory (see \link[base]{getwd}) unless the full path is specified.}

\item{VNC}{Object of class \linkS4class{VarNameCorresp}.}

\item{DDslot}{Character vector of length 1 with the name of DD slot in which transformation will 
be made. Its default value is \code{NULL}. If no DDslot is specified, variables in VNC components 
are assigned to the corresponding DD slot.}
}
\value{
Object of class \linkS4class{DD}.
}
\description{
This function is a constructor for the class \linkS4class{DD} using the contents of 
a xml file.

\code{RepoDDToDD} read xml files with the definition and properties of every variable and
transform this content into an object of class \linkS4class{DD}. 

This function internally builds a \linkS4class{data.table} with columns \code{Variable}, 
\code{Sort}, \code{Class}, \code{Qual1} to \code{Qual}\emph{q} and ValueRegExp.
 
The column \code{Variable} contains the names of all variables, both questionnaire variables and 
metadata. This internal \linkS4class{data.table} is then used to initialize a \linkS4class{DD} 
object.

The column \code{Sort} takes values \code{'IDQual'}, \code{'NonIDQual'} or \code{'IDDD'}, for 
statistical unit qualifiers, variable name qualifiers and variable names, respectively.

The column \code{Class} specifies the class of the variable and takes values \code{numeric} or 
\code{character}. 

The column \code{Length} contains the highest length for each variable.

The columns \code{Qual1} to \code{Qual}\emph{q} contain the names of the qualifiers of every 
variable name (row).

The column \code{ValueRegExp} contains an expresion with the acceptable values for each variable.
}
\examples{
\dontrun{
# We assume that the DD file (xml file) ExampleXLS.DD_V1 already exists in the administrator 
# desktop (change accordingly otherwise):
DDFileName <- 'C:/Users/Administrador/Desktop/ExampleXLS.DD_V1'
data(ExampleVNC)
RepoDDToDD(DDFileName, ExampleVNC)
}

}

