% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoFileToStQList.R
\name{RepoFileToStQList}
\alias{RepoFileToStQList}
\title{Produce an object of class \code{\link[StQ]{StQList}} from files with key-value pair structure}
\usage{
RepoFileToStQList(
  SurveyCode,
  RepoPath,
  FileType,
  IniPeriod,
  FinPeriod,
  Base,
  Rot = FALSE,
  includeFI = TRUE,
  perl = FALSE,
  sep = "@@",
  encoding = "unknown"
)
}
\arguments{
\item{SurveyCode}{Character vector of length 1 with the code of each survey.}

\item{RepoPath}{Character vector of length 1 with the path of the repository from which files are
to be read.}

\item{FileType}{Character vector of length 1 with the type of the file to be read (FI, FP, FF, FG
FD, FL or FT).}

\item{IniPeriod}{Character vector of length 1 with the initial time period to be read (in the 
repository notation).}

\item{FinPeriod}{Character vector of length 1 with the final time period to be read (in the 
repository notation).}

\item{Base}{Character vector of length 1 with the year ('aaaa') of the base to which data are
referred. If it has no sense, "Base" is a empty character vector.}

\item{Rot}{Logical vector of length 1 indicating whether rotated sample files are to be included 
(default value FALSE).}

\item{includeFI}{Logical vector of length 1 indicating whether ID variables will be included in
the  slot of class \code{\link[StQ]{VNC}} of the \link{DD} object.}

\item{perl}{Logical vector of length 1 indicating whether Perl is installed in the system or not.}

\item{sep}{Logical vector of length 1 containing the combination of characters used as separator 
in the input file (default value @).}

\item{encoding}{Default value is "unknown". Other possible options are "UTF-8" and "Latin-1". 
Note: it is not used to re-encode the input, rather enables handling of encoded strings in their 
native encoding.}
}
\value{
Object of class \code{\link[StQ]{StQList}}.
}
\description{
\code{RepoFileToStQList} returns an object of class \code{\link[StQ]{StQList}} from files
with key-value pair structure.
}
\examples{
\dontrun{
 RepoFileToStQList('E30163', 'T:/E30163/', 'FF', 'MM022016', 'MM032016', '', perl = TRUE)
}

}
