% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoXLSToRepoDD.R
\name{RepoXLSToRepoDD}
\alias{RepoXLSToRepoDD}
\title{Produce a file DD from an xlsx file}
\usage{
RepoXLSToRepoDD(SurveyCode, Version)
}
\arguments{
\item{SurveyCode}{Character vector of length 1 with the code of the survey. The xlsx file will be 
read from the working directory (see \link[base]{getwd}) unless the full path is specified.}

\item{Version}{Character vector of length 1 with the version number of the file DD to produce.}
}
\value{
NULL.
}
\description{
This function builds and writes a DD file  using the contents of an xlsx file.

\code{RepoXLSToRepoDD} transforms the content of an xlsx file into a DD file.
}
\examples{
\dontrun{
# We assume that the xlsx file \\code{E30163.NombresVariables.xlsx} with the appropriate structure
# is in the administrator desktop (change accordingly otherwise):
RepoXLSToRepoDD(SurveyCode, Version)
}

}

