% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepoXLSToRepoDD.R
\name{RepoXLSToRepoDD}
\alias{RepoXLSToRepoDD}
\title{Produce a file DD from an xlsx file}
\usage{
RepoXLSToRepoDD(ExcelName)
}
\arguments{
\item{ExcelName}{Character vector of length 1 with the name of the file to read.The file will be 
read from the working directory (see \link[base]{getwd}) unless the full path is specified.}
}
\value{
Return invisible NULL.
}
\description{
This function builds and writes a DD file  using the contents of an xlsx file.
\code{RepoXLSToRepoDD} transforms the content of an xlsx file into a DD file.
}
\examples{
# We assume that the xlsx file ExampleXLS.NombresVariables.xlsx with the appropriate structure is
# in the working directory (change accordingly otherwise):
\dontrun{
RepoDD <- RepoXLSToRepoDD('ExampleXLS')
show(RepoDD)
}

}

