% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlToDD.R
\name{xmlToDD}
\alias{xmlToDD}
\title{Transforma el fichero xml del repositorio a su formato para R}
\usage{
xmlToDD(FileName)
}
\arguments{
\item{FileName}{\code{\link{vector}} de tipo \code{character} de longitud 1 
con el nombre del fichero que se desea leer. El nombre debe incluir la ruta
completa de ubicación del fichero (directorio en el que se encuentra el
fichero y su nombre).}
}
\value{
Objeto de clase \code{\linkS4class{DD}}, en el formato adecuado para
su inclusión como slot \code{DD} de un objeto de clase \code{StQ}.
}
\description{
\code{xmlToDD} lee los ficheros xml ubicados en el repositorio de microdatos 
de la encuesta y transforma el contenido en un \code{\linkS4class{data.table}}
con el formato requerido para su inclusión como slot \code{DD} de un objeto
de clase \code{\linkS4class{StQ}}.
}
\details{
Esta función toma el contenido del fichero xml de entrada y 
construye un \code{data.table} con las columnas \code{Variable}, 
\code{Sort}, \code{Class} and \code{Qual\code{1}} to \code{Qual\code{q}}.

La columna \code{Variable} contiene los nombres de todas las variables, tanto
de las del cuestionario como de las creadas durante el proceso.

La columna \code{Sort} toma los valores \code{'IDQual'}, \code{'NonIDQual'}, 
e \code{'IDDD'}, para calificadores de identificación de las unidades 
estadísticas, para calificadores de nombres de variables y para nombres de
variables otras variables en el cuestionario.

La columna \code{Class} especifica la clase de la variable, y toma los valores
\code{numeric} o \code{character}.

Las columnas \code{Qual\code{1}} a \code{Qual\code{q}} contienen los nombres de los 
calificadores de cada variable (fila).

Se transforma el fichero de definición de datos para que su tratamiento desde
el punto de vista de la programación sea más eficiente.
}
\examples{

FileName<-"N:/UDMTD/UDTMDCOM/DepSel.SoftwareR/INE.Packages.v2/Ejemplo_v2.xml"
DD <- xmlToDD(FileName)

#@include DD-class.R

}

